refactor: add unit tests for Converters class to validate List<Long> serialization
